api/pyproject.toml
[project]
name = "wcs-api"
version = "0.1.0"
description = "SwingFlow analysis API (music + video)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic-settings>=2.6.0",
  "python-multipart>=0.0.12",
  "pyjwt>=2.9.0",
  "librosa>=0.10.2",
  "numpy>=1.26.0",
  "soundfile>=0.12.1",
  "audioread>=3.0.1",
  "stripe>=11.0.0",
  "httpx>=0.27.0",
  "google-genai>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wcs_api"]
